Restructure plugin package. Separate connections and directives
