+++ sphinx-asdf/directives.py
@@ -1,27 +1,22 @@
-import os
 import posixpath
 from pprint import pformat
 
 import yaml
 
 from docutils import nodes
-from docutils.frontend import OptionParser
 from docutils.statemachine import ViewList
 
 from sphinx import addnodes
-from sphinx.parsers import RSTParser
-from sphinx.util.fileutil import copy_asset
 from sphinx.util.nodes import nested_parse_with_titles
-from sphinx.util.docutils import SphinxDirective, new_document
+from sphinx.util.docutils import SphinxDirective
 
 from .md2rst import md2rst
-from .nodes import (add_asdf_nodes, toc_link, schema_header_title,
-                    schema_title, schema_description, schema_properties,
-                    schema_property, schema_property_name,
-                    schema_property_details, schema_combiner_body,
-                    schema_combiner_list, schema_combiner_item, section_header,
-                    asdf_tree, asdf_ref, example_section, example_item,
-                    example_description)
+from .nodes import (toc_link, schema_header_title, schema_title,
+                    schema_description, schema_properties, schema_property,
+                    schema_property_name, schema_property_details,
+                    schema_combiner_body, schema_combiner_list,
+                    schema_combiner_item, section_header, asdf_tree, asdf_ref,
+                    example_section, example_item, example_description)
 
 
 SCHEMA_DEF_SECTION_TITLE = 'Schema Definitions'
@@ -372,127 +367,3 @@ class AsdfSchema(SphinxDirective):
             node.append(nodes.literal_block(text=example[1], language='yaml'))
             examples.append(node)
         return examples
-
-
-def find_autoasdf_directives(env, filename):
-
-    parser = RSTParser()
-    settings = OptionParser(components=(RSTParser,)).get_default_values()
-    settings.env = env
-    document = new_document(filename, settings)
-
-    with open(filename) as ff:
-        parser.parse(ff.read(), document)
-
-    return [x.children[0].astext() for x in document.traverse()
-            if isinstance(x, schema_def)]
-
-
-def find_autoschema_references(app, genfiles):
-
-    # We set this environment variable to indicate that the AsdfSchemas
-    # directive should be parsed as a simple list of schema references
-    # rather than as the toctree that will be generated when the documentation
-    # is actually built.
-    app.env.autoasdf_generate = True
-
-    schemas = set()
-    for fn in genfiles:
-        # Create documentation files based on contents of asdf-schema directives
-        path = posixpath.join(app.env.srcdir, fn)
-        app.env.temp_data['docname'] = app.env.path2doc(path)
-        schemas = schemas.union(find_autoasdf_directives(app.env, path))
-
-    # Unset this variable now that we're done.
-    app.env.autoasdf_generate = False
-
-    return list(schemas)
-
-
-def create_schema_docs(app, schemas):
-
-    standard_prefix = app.env.config.asdf_schema_standard_prefix
-    output_dir = posixpath.join(app.srcdir, 'generated', standard_prefix)
-    os.makedirs(output_dir, exist_ok=True)
-
-    for schema_name in schemas:
-        doc_path = posixpath.join(output_dir, schema_name + '.rst')
-
-        if posixpath.exists(doc_path):
-            continue
-
-        os.makedirs(posixpath.dirname(doc_path), exist_ok=True)
-
-        with open(doc_path, 'w') as ff:
-            ff.write(schema_name + '\n')
-            ff.write('=' * len(schema_name) + '\n\n')
-            ff.write('.. asdf-schema::\n\n')
-            ff.write('    {}\n'.format(schema_name))
-
-
-def autogenerate_schema_docs(app):
-
-    env = app.env
-
-    genfiles = [env.doc2path(x, base=None) for x in env.found_docs
-                if posixpath.isfile(env.doc2path(x))]
-
-    if not genfiles:
-        return
-
-    ext = list(app.config.source_suffix)
-    genfiles = [genfile + (not genfile.endswith(tuple(ext)) and ext[0] or '')
-                for genfile in genfiles]
-
-    # Read all source documentation files and parse all asdf-schema directives
-    schemas = find_autoschema_references(app, genfiles)
-    # Create the documentation files that correspond to the schemas listed
-    create_schema_docs(app, schemas)
-
-
-def add_labels_to_nodes(app, document):
-    labels = app.env.domaindata['std']['labels']
-    anonlabels = app.env.domaindata['std']['anonlabels']
-    basepath = os.path.join('generated', app.env.config.asdf_schema_standard_prefix)
-
-    for node in document.traverse():
-        if isinstance(node, str) or not (isinstance(node, nodes.Node) and node['ids']):
-            continue
-
-        labelid = node['ids'][0]
-        docname = app.env.docname
-        basename = os.path.relpath(docname, basepath)
-        name = nodes.fully_normalize_name(basename + ':' + labelid)
-
-        # labelname -> docname, labelid
-        anonlabels[name] = docname, labelid
-        # labelname -> docname, labelid, sectionname
-        labels[name] = docname, labelid, ''
-
-
-def on_build_finished(app, exc):
-    if exc is None:
-        for asset in ['sphinx_asdf.css', 'sphinx_asdf.js']:
-            src = posixpath.join(posixpath.dirname(__file__), 'static', asset)
-            dst = posixpath.join(app.outdir, '_static')
-            copy_asset(src, dst)
-
-
-def setup(app):
-
-    # Describes a path relative to the sphinx source directory
-    app.add_config_value('asdf_schema_path', 'schemas', 'env')
-    app.add_config_value('asdf_schema_standard_prefix', '', 'env')
-    app.add_directive('asdf-autoschemas', AsdfSchemas)
-    app.add_directive('asdf-schema', AsdfSchema)
-
-    add_asdf_nodes(app)
-
-    app.add_css_file('sphinx_asdf.css')
-    app.add_javascript('sphinx_asdf.js')
-
-    app.connect('builder-inited', autogenerate_schema_docs)
-    app.connect('doctree-read', add_labels_to_nodes)
-    app.connect('build-finished', on_build_finished)
-
-    return dict(version='0.1')
